requirements.txt
Django==2.2.11
# --- pytz==2019.3
# --- sqlparse==0.3.1
psycopg2-binary==2.8.4
coverage==5.0.4
celery==4.3.0
# --- amqp==2.5.2
# --- billiard==3.6.3.0
# --- importlib-metadata==1.6.0
# --- kombu==4.6.8
# --- vine==1.3.0
# --- zipp==3.1.0
redis==3.4.1
PyYAML==5.3.1
djangorestframework==3.11.0
